+++ src/pages/Shop/Shoes/AllShoes.jsx
@@ -1,18 +1,18 @@
 import styled from 'styled-components';
-import Item from '../Item';
-import Nomination from '../Nomination';
-import ShopHeader from '../ShopHeader';
+import Item from './Item';
+import Nomination from './Nomination';
+import ShopHeader from './ShopHeader';
 
 const StyledAllShoes = styled.div``;
 
 const AllShoes = () => {
-	return (
-		<StyledAllShoes>
-			<ShopHeader />
-			<Nomination />
-			<Item />
-		</StyledAllShoes>
-	);
+  return (
+    <StyledAllShoes>
+      <ShopHeader />
+      <Nomination />
+      <Item />
+    </StyledAllShoes>
+  );
 };
 
 export default AllShoes;
