--- src/pages/Shop/ShopMain.jsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-import Item from './Item';
-
-const StyledShopMain = styled.div``;
-
-const ShopMain = () => {
-  return (
-    <StyledShopMain>
-      <Item />
-    </StyledShopMain>
-  );
-};
-
-export default ShopMain;
